+++ Assets/Scripts/Services/MonoBehaviourWithBarDependency.cs
@@ -1,0 +1,12 @@
+using UnityEngine;
+
+namespace MrWatts.Internal.FuelInject.TestProject
+{
+    public sealed class MonoBehaviourWithBarDependency : MonoBehaviour
+    {
+        [Inject]
+        private Bar Bar { get; set; } = default!;
+
+        public Bar BarGetter { get => Bar; }
+    }
+}