+++ Assets/Scripts/Services/MonoBehaviourWithIInjectorSceneDependency.cs
@@ -1,0 +1,13 @@
+using UnityEngine;
+using UnityEngine.SceneManagement;
+
+namespace MrWatts.Internal.FuelInject.TestProject
+{
+    public sealed class MonoBehaviourWithIInjectorSceneDependency : MonoBehaviour
+    {
+        [Inject]
+        private IInjector<Scene> Injector { get; set; } = default!;
+
+        public IInjector<Scene> InjectorGetter { get => Injector; }
+    }
+}