+++ Assets/Scripts/TestSceneModule.cs
@@ -6,14 +6,11 @@ namespace MrWatts.Internal.FuelInject.TestProject
     public sealed class TestSceneModule : MonoBehaviour, IUnityContainerModule
     {
         [SerializeField]
-        private FooMonoBehaviour fooMonoBehaviour = default!;
+        private MonoBehaviourWithBarDependency monoBehaviourWithBarDependency = default!;
 
         public void Configure(ContainerBuilder builder)
         {
-            // TODO: Add proper tests.
-            Debug.Log("Test module configuring container");
-
-            builder.RegisterInstance(fooMonoBehaviour);
+            builder.RegisterInstance(monoBehaviourWithBarDependency);
 
             builder.RegisterType<Foo>().SingleInstance();
             builder.RegisterType<Bar>().SingleInstance();