+++ Assets/Package/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 -   Internal restructuring.
 -   `IInjector<object>`, `IInjector<GameObject>`, and `IInjector<Scene>` are now available for clients to use when they need to inject dependencies into objects at runtime (e.g. after the scene was already loaded).
+-   `SceneTestGroup` in assembly `mrwatts.fuelinject.testing` can now be inherited from to make scene testing easier.
 
 # 0.1.1